+++ src/renderer/prepare.rs
@@ -1,12 +1,11 @@
+use super::extract::{ExtractedPixelScale, ExtractedRenderText, ExtractedRenderVector};
+use crate::{assets::vector::VelloVector, color_swapping::ColorPaletteSwap, RenderMode, Vector};
 use bevy::{
     prelude::*,
     render::{camera::ExtractedCamera, render_asset::RenderAssets, view::ExtractedView},
 };
 use vello::kurbo::Affine;
-
-use crate::{assets::vector::VelloVector, RenderMode};
-
-use super::extract::{ExtractedPixelScale, ExtractedRenderText, ExtractedRenderVector};
+use vellottie::runtime::model::Brush;
 
 #[derive(Component, Copy, Clone)]
 pub struct PreparedAffine(pub Affine);
@@ -157,5 +156,105 @@ pub fn prepare_text_affines(
         commands
             .entity(entity)
             .insert(PreparedAffine(Affine::new(transform)));
+    }
+}
+
+pub fn prepare_vector_composition_edits(mut render_vectors: Query<&mut ExtractedRenderVector>) {
+    // A depth-first traversal and remap of colors, O(n)
+    'vectors: for mut render_vector in render_vectors.iter_mut() {
+        // Get vector and color swap or there's no use continuing...
+        let ExtractedRenderVector {
+            render_data,
+            color_swaps,
+            ..
+        } = render_vector.as_mut();
+
+        // Continue if there are no colors
+        let Some(ColorPaletteSwap { ref colors }) = color_swaps else {
+            continue 'vectors;
+        };
+
+        // Perform recolors!
+        // TODO: Recoloring SVGs
+        let Vector::Animated(ref mut composition) = render_data.data else {
+            continue 'vectors;
+        };
+        'layers: for layer in composition.layers.iter_mut() {
+            // Continue if this layer doesn't have a color swap
+            let Some(target_color) = colors.get(&layer.name) else {
+                continue 'layers;
+            };
+            let shapes = match &mut layer.content {
+                vellottie::runtime::model::Content::Shape(shapes) => shapes,
+                vellottie::runtime::model::Content::None
+                | vellottie::runtime::model::Content::Instance { .. } => {
+                    continue 'layers;
+                }
+            };
+            let target_color = vello::peniko::Color::rgba(
+                target_color.r().into(),
+                target_color.g().into(),
+                target_color.b().into(),
+                target_color.a().into(),
+            );
+            for shape in shapes.iter_mut() {
+                match shape {
+                    vellottie::runtime::model::Shape::Group(shapes, _) => {
+                        for shape in shapes.iter_mut() {
+                            let vellottie::runtime::model::Shape::Draw(ref mut draw) = shape else {
+                                continue;
+                            };
+                            recolor_brush(&mut draw.brush, target_color);
+                        }
+                    }
+                    vellottie::runtime::model::Shape::Draw(draw) => {
+                        recolor_brush(&mut draw.brush, target_color);
+                    }
+                    vellottie::runtime::model::Shape::Repeater(_)
+                    | vellottie::runtime::model::Shape::Geometry(_) => {
+                        continue;
+                    }
+                }
+            }
+        }
+    }
+}
+
+/// A helper method  to recolor a brush with a target color.
+fn recolor_brush(brush: &mut Brush, target_color: vello::peniko::Color) {
+    match brush {
+        vellottie::runtime::model::Brush::Fixed(brush) => match brush {
+            vello::peniko::Brush::Solid(solid) => {
+                *solid = target_color;
+            }
+            vello::peniko::Brush::Gradient(gradient) => {
+                for stop in gradient.stops.iter_mut() {
+                    stop.color = target_color;
+                }
+            }
+            vello::peniko::Brush::Image(_) => {}
+        },
+        vellottie::runtime::model::Brush::Animated(brush) => match brush {
+            vellottie::runtime::model::animated::Brush::Solid(brush) => match brush {
+                vellottie::runtime::model::Value::Fixed(solid) => {
+                    *solid = target_color;
+                }
+                vellottie::runtime::model::Value::Animated(keyframes) => {
+                    for solid in keyframes.values.iter_mut() {
+                        *solid = target_color;
+                    }
+                }
+            },
+            vellottie::runtime::model::animated::Brush::Gradient(gr) => match &mut gr.stops {
+                vellottie::runtime::model::ColorStops::Fixed(stops) => {
+                    for stop in stops.iter_mut() {
+                        stop.color = target_color;
+                    }
+                }
+                vellottie::runtime::model::ColorStops::Animated(_stops) => {
+                    // FIXME: Why does stops use f32 instead of color?
+                }
+            },
+        },
     }
 }